Video Player Maker update - v0.6.1
This patch addresses an issue that made it impossible to set an automatic framerate limit (using 0 as the FPS in TurboWarp). (Did you know you can do that?) It's also very easy for creators to change the behavior, under the "step frame" custom block definition, like always using delta time.

- When the framerate limit is set to 0 (which means automatic) in TurboWarp, delta time is calculated for all over-time events, including the video itself, instead of just referencing the framerate limit. Otherwise, everything continues to behave as it did before.

Resolves #5